requirements.txt
# Web Framework
flask>=3.0.0

# Environment Configuration
python-dotenv>=1.0.0

# HTTP Requests (GitHub API)
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0